Fix  deletebyid for vacancy and candidatedocumet

+++ Android/Interview/app/src/main/java/com/example/interview/models/responses/get/session/SessionResponse.kt
@@ -10,7 +10,7 @@ data class SessionResponse(
     @SerializedName("id")
     val id: Int?,
     @SerializedName("endValue")
-    val endValue: Int?,
+    val endValue: Double?,
     @SerializedName("startDate")
     val startDate: String?,
     @SerializedName("endDate")
